speedrun.cpp: make it finally work!
This was a hard task reimplement all the task in a single file without
anything beyond the standard library, but it seems like I did it.
Now I want my cookies.

However, the result is a quite ad-hoc solutions with many quirks and
shortcuts. Production code is not supposed to be written this way.

Signed-off-by: Pavel Kretov <firegurafiku@gmail.com>